WIP(backtest): refactor
